@@ -1,9 +1,11 @@
 module github.com/vinyldns/go-vinyldns
 
-go 1.17
+go 1.21
 
 require (
+	github.com/aws/aws-sdk-go-v2 v1.26.1
+	github.com/aws/aws-sdk-go-v2/credentials v1.17.11
 	github.com/gobs/pretty v0.0.0-20180724170744-09732c25a95b
-	github.com/smartystreets/go-aws-auth v0.0.0-20180515143844-0c1422d1fdb9
-	github.com/smartystreets/gunit v1.0.4 // indirect
 )
+
+require github.com/aws/smithy-go v1.20.2 // indirect
